libcoveweb2: Pass SuppliedDataFiles to tasks
This stops tasks having to reload the same data multiple times.

It also means that processes can set temporary info on a SuppliedDataFile
and be sure that a later task in the same process pipeline will have
access to it.
This is not currently used but may be used in future.

Also renamed supplied_data_json_files var in ConvertSpreadsheetIntoJSON.
It's not json so the name is misleading.