libcoveweb2: Add meta to SuppliedData
In OFDS meta on SuppliedDataFile is used to say
which file is nodes and which is spans.

In BODS meta on SuppliedData will be used to say
whether it's in sample mode or not,
and it doesn't make sense for that to be on a file.

Change meta to always have a dict so code doesn't need to check for null's.
In effect, because the 3 save methods on SuppliedData have meta={} null
was never set anyway.